renames contact-us to contact
