+++ frontend/src/components/footer.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 
 const navigation = [
   { name: 'About', href: '/about' },
-  { name: 'Contact', href: '/contact-us' },
+  { name: 'Contact', href: '/contact' },
   { name: 'Privacy Policy', href: '/privacy-policy' },
 ];
