+++ frontend/src/pages/api/contact.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { NextApiRequest, NextApiResponse } from 'next';
 
-import { ContactUsInput } from '@/containers/contact-us/form';
+import { ContactUsInput } from '@/containers/contact/form';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === 'POST') {
